Load synthetics as a plugin
