+++ packages/datadog-ci/src/cli.ts
@@ -1,10 +1,14 @@
 import fs from 'fs'
 
+import {baseCommands} from '@datadog/datadog-ci-base/cli'
 import {Builtins, Cli, CommandClass} from 'clipanion'
 
 import {cliVersion} from './version'
 
-export const BETA_COMMANDS = ['dora', 'deployment', 'elf-symbols']
+export const BETA_COMMANDS = new Set(['dora', 'deployment', 'elf-symbols'])
+
+const betaCommandsEnabled =
+  process.env.DD_BETA_COMMANDS_ENABLED === '1' || process.env.DD_BETA_COMMANDS_ENABLED === 'true'
 
 const onError = (err: any) => {
   console.log(err)
@@ -23,19 +27,28 @@ const cli = new Cli({
 cli.register(Builtins.HelpCommand)
 cli.register(Builtins.VersionCommand)
 
+// Commands not migrated yet
 const commandsPath = `${__dirname}/commands`
 for (const commandFolder of fs.readdirSync(commandsPath)) {
-  const betaCommandsEnabled =
-    process.env.DD_BETA_COMMANDS_ENABLED === '1' || process.env.DD_BETA_COMMANDS_ENABLED === 'true'
-  if (BETA_COMMANDS.includes(commandFolder) && !betaCommandsEnabled) {
+  if (!betaCommandsEnabled && BETA_COMMANDS.has(commandFolder)) {
     continue
   }
+
   const commandPath = `${commandsPath}/${commandFolder}`
   if (fs.statSync(commandPath).isDirectory()) {
     // eslint-disable-next-line @typescript-eslint/no-var-requires
     ;(require(`${commandPath}/cli`) as CommandClass[]).forEach((command) => cli.register(command))
   }
 }
+
+// Commands migrated to the `@datadog/datadog-ci-base` package
+Object.entries(baseCommands).forEach(([scope, commands]) => {
+  if (!betaCommandsEnabled && BETA_COMMANDS.has(scope)) {
+    return
+  }
+
+  commands.forEach((command) => cli.register(command))
+})
 
 if (require.main === module) {
   void cli.runExit(process.argv.slice(2), {
