+++ packages/plugin-synthetics/src/cli.ts
@@ -1,6 +1,6 @@
+import {PluginCommand as RunTestsCommand} from './commands/run-tests'
 import {DeployTestsCommand} from './deploy-tests-command'
 import {ImportTestsCommand} from './import-tests-command'
-import {RunTestsCommand} from './run-tests-command'
 import {UploadApplicationCommand} from './upload-application-command'
 
 module.exports = [RunTestsCommand, UploadApplicationCommand, ImportTestsCommand, DeployTestsCommand]
